% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{card}
\alias{card}
\title{Generate a single card}
\usage{
card(
  img,
  name,
  job_level,
  highlights_1,
  highlights_2,
  highlights_3,
  dev_1,
  dev_2,
  dev_3,
  months_in_role,
  rating,
  job_colour,
  promotion,
  months_in_role_cuts
)
}
\value{
Tibble with info per-person. I
}
\description{
Generate a single card
}
