% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_people.R
\name{enrich_people}
\alias{enrich_people}
\title{Merge the people and config files}
\usage{
enrich_people(
  people = NULL,
  config = NULL,
  months_in_role_cuts = c(12 * 3, 12 * 6)
)
}
\arguments{
\item{people}{People from the output of [read_people()].}

\item{config}{Config data from [read_config()].}

\item{months_in_role_cuts}{Numeric vector of length 2 with cut points for time in role badges.}
}
\value{
Tibble with info per-person. I
}
\description{
`enrich_people()` Merge the people and config file outputs from the `read_`
files.
}
