% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Read in the formatted 'config' YAML}
\usage{
read_config(path = ".", file = "config.yaml")
}
\arguments{
\item{path}{The path to the file of interest. Defaults to current dir.}

\item{file}{This is likely to always be \code{config.yaml}, but I was stuck
between calling it config, and 'lookups' - as initially it only contains
lookups.
See the function \code{\link[=enrich_people]{enrich_people()}} to tidy up the data.}
}
\description{
\code{read_config()} pulls in the config file, that has info like what
are the different roles, and allowable values for some parameters (for CICD).
}
